Move database.js out of store directory to keep it from being seen as a store module.
