+++ store/index.js
@@ -1,5 +1,5 @@
 import VuexORM from '@vuex-orm/core'
-import database from './database'
+import database from '../orm/database'
 
 export const plugins = [ VuexORM.install(database) ]
